consolidate Token and TokenInRoute types. move some types to flame-types.
