+++ packages/flame-types/package.json
@@ -14,5 +14,8 @@
     "@repo/eslint-config": "*",
     "@repo/typescript-config": "*",
     "typescript": "5.5.4"
+  },
+  "dependencies": {
+    "jsbi": "^4.3.0"
   }
 }
